Made a couple changes
I should've made a lot more commits along the way. Main changes I made is just adding in the projectiles and modifying enemy so that they can fire the projectiles and check for collisions. I still need to add in the randomized part.

Other than that, I just modified the player class so that it inherits from Sprite instead of having a Sprite data member, because it makes it easier to get the global bounds when I'm checking for a collision in the app folder. Will have another commit before I make a pull request