Projectile shooting making me mad
There are some slight bugs that I can't seem to figure out, but they're not stopping the game from function as intended for the most part.

Main thing is when you get hit, when everything resets, the number of balls spawned in is increased for some reason. I had though that it was just due to the fact that I hadn't deleted all of the the balls but even when I did that they still spawned at an abnormal rate. If anyone could look it over to see if they spot the issue that'd be appreciated